plugins: add flag to 'download:plugins' script
Fixes #7122

Adds the following flag to the `download:plugins` script
to control the behavior of how plugins should be packed/unpacked by
default.

The default is `false` and it means that plugins should not be preserved
as `.vsix` and instead should be unpacked. If `preserveVsix` is set to `true`
then plugins are preserved in their original format.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>